requirements.txt
python-dotenv==1.0.1
langchain-core==0.3.45
pypdf==5.3.1
google-genai==1.5.0
qdrant-client==1.13.3
fastembed==0.6.0
sentence-transformers==3.4.1
python-docx==1.1.2
langchain==0.3.20
streamlit==1.43.2
langchain_google_genai==2.1.0
pandas==2.0.3
scikit-learn==1.3.2
cryptography==44.0.2
